Moved function export_image to exportutils module
